Actual voltage is passed.
Changed the voltage methods so it passes an actual voltage instead of a value between 0-1023. The value is a float, so the floatToByte method was used instead of the intToByte method. 
The intToByte method still has an issue. If the decimal value being translated ends up having a value of zero for one of the hex bytes, the zero is discarded. This warps the value of the number.
The intToByte method needs to be fixed.