Fixed inaccuracy caused by torqueOff().
The method servo.torqueOff() from the XYZrobotServo.h library is called in the noTorque() method. 
Their .torqueOff() method does not update.
In the method torqueControl(), the current speed of the motor is now read in using the getOmegaActual() method. More reliable than the byteToInt() method call, gives the wrong value for negative numbers and the first value returned by the method is also a large value.